+++ flourish_metadata_rules/metadata_rules/caregiver_rule_groups/subject_visit_rules.py
@@ -156,6 +156,13 @@ class MaternalVisitRuleGroup(CrfRuleGroup):
         alternative=NOT_REQUIRED,
         target_models=[f'{app_label}.parentadolrelationshipscale', ])
 
+    childhood_lead_exposure_risk = CrfRule(
+        predicate=pc.func_childhood_lead_exposure_risk_required,
+        consequence=REQUIRED,
+        alternative=NOT_REQUIRED,
+        target_models=[f'{app_label}.childhoodleadexposurerisk', ]
+    )
+
     class Meta:
         app_label = app_label
         source_model = f'{app_label}.maternalvisit'
