+++ flourish_metadata_rules/metadata_rules/child_rule_groups/child_visit_rules.py
@@ -149,13 +149,6 @@ class ChildVisitRuleGroup(CrfRuleGroup):
         alternative=NOT_REQUIRED,
         target_models=[f'{app_label}.childsafistigma', ])
 
-    childhood_lead_exposure_risk = CrfRule(
-        predicate=pc.func_childhood_lead_exposure_risk_required,
-        consequence=REQUIRED,
-        alternative=NOT_REQUIRED,
-        target_models=[f'{app_label}.childhoodleadexposurerisk', ]
-    )
-
     class Meta:
         app_label = app_label
         source_model = f'{app_label}.childvisit'
