+++ flourish_metadata_rules/metadata_rules/child_rule_groups/infant_hiv_test_rules.py
@@ -35,8 +35,6 @@ class InfantHIVTestRuleGroup(CrfRuleGroup):
 
     hiv_testing_6to8_months = CrfRule(
         predicate=pc.hiv_test_6_to_8_weeks_required,
-    birth_hiv_testing = CrfRule(
-        predicate=pc.hiv_test_birth_required,
         consequence=REQUIRED,
         alternative=NOT_REQUIRED,
         target_models=[f'{app_label}.infanthivtestingage6to8weeks', ])
