+++ flourish_metadata_rules/predicates/child_predicates.py
@@ -1,8 +1,5 @@
 from datetime import timedelta
 
-import pytz
-from dateutil.relativedelta import relativedelta
-
 from django.apps import apps as django_apps
 from django.db.models import Q
 from edc_base.utils import age, get_utcnow
@@ -727,22 +724,6 @@ class ChildPredicates(PredicateCollection):
 
             )
         return False
-
-    def func_childhood_lead_exposure_risk_required(self, visit=None, **kwargs):
-        childhood_lead_exposure_risk_model = f'{self.app_label}.childhoodleadexposurerisk'
-        prev_instance = self.previous_model(visit=visit,
-                                            model=childhood_lead_exposure_risk_model)
-        is_follow_up = '300' in visit.visit_code
-
-        if prev_instance:
-            visit_definition = visit.appointment.visits.get(visit.appointment.visit_code)
-            earlist_appt_date = (visit.appointment.timepoint_datetime -
-                                 visit_definition.rlower).astimezone(
-                pytz.timezone('Africa/Gaborone'))
-            return (earlist_appt_date - prev_instance.report_datetime) > timedelta(
-                days=365)
-        else:
-            return is_follow_up
 
     def hiv_test_required(self, child_age, visit):
         try:
